requirements.txt
openai
whisper
requests
pandas
python-dotenv
git+https://github.com/openai/whisper.git
torch  --index-url https://download.pytorch.org/whl/cu118
torchvision  --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118
